move parseUnit to VsStepper
