+++ packages/vlossom/src/utils/string.ts
@@ -13,28 +13,4 @@ export const stringUtil = {
             .map((s) => s.toLowerCase())
             .join('-');
     },
-
-    parseUnit(str: string | number): { value: number; unit: string } {
-        const numValue = Number(str);
-
-        if (typeof str === 'number' || !isNaN(numValue)) {
-            return {
-                value: numValue,
-                unit: 'px',
-            };
-        }
-
-        const match = str.match(/(\d+)(\w+|%)/);
-
-        if (match) {
-            return {
-                value: Number(match[1]),
-                unit: match[2],
-            };
-        }
-        return {
-            value: 0,
-            unit: '',
-        };
-    },
 };
